+++ tensorflow_quantum/core/src/util_balance_trajectory.cc
@@ -75,9 +75,9 @@ void BalanceTrajectory(const int& num_samples, const int& num_threads,
 
   int prev_max_height = -1;
   for (size_t j = 0; j < (*thread_offsets)[0].size(); j++) {
-    int run_ceiling = ((num_samples + num_threads - 1) / num_threads);
-    int num_lo = num_threads * run_ceiling - num_samples;
-    int num_hi = num_threads - num_lo;
+    uint64_t run_ceiling = ((static_cast<uint64_t>(num_samples) + num_threads - 1) / num_threads);
+    uint64_t num_lo = num_threads * run_ceiling - num_samples;
+    uint64_t num_hi = num_threads - num_lo;
     int cur_max = prev_max_height;
     for (int i = 0; i < num_threads; i++) {
       if (height[i] == cur_max && num_lo) {
